backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2

# Authentication
python-jose[cryptography]==3.3.0
httpx==0.27.2

# Configuration
pydantic==2.9.2
pydantic-settings==2.5.2

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0

# Development
black==24.8.0
isort==5.13.2
mypy==1.11.2
